Airflow Lab/requirements.txt
pandas
scikit-learn
numpy
matplotlib
apache-airflow